implemented global exception for handling not found exception for course, grade , student.

+++ src/main/java/com/gradems/grademangementsystem/service/CourseServiceImpl.java
@@ -1,11 +1,13 @@
 package com.gradems.grademangementsystem.service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
 import com.gradems.grademangementsystem.entity.Course;
+import com.gradems.grademangementsystem.exception.NotFoundException;
 import com.gradems.grademangementsystem.repository.CourseRepository;
 
 import lombok.AllArgsConstructor;
@@ -18,22 +20,33 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public Course getCourse(UUID id) {
-        return courseRepository.findById(id).get();
+        Optional<Course> course =  courseRepository.findById(id);
+        return unwrapCourse(course, id);
     }
 
     @Override
     public Course saveCourse(Course course) {
+        //TODO validation of entity or exception to be passed with proper status code
         return courseRepository.save(course);
     }
 
     @Override
     public void deleteCourse(UUID id) {
-        courseRepository.deleteById(id);
+        if(getCourse(id) !=  null) {
+            courseRepository.deleteById(id);
+        }
     }
 
     @Override
     public List<Course> getCourses() {
         return (List<Course>)courseRepository.findAll();
     }
-    
+
+    static Course unwrapCourse(Optional<Course> course, UUID id) {
+        if(course.isPresent()){
+            return course.get();
+        } else {
+            throw new NotFoundException(id);
+        }
+    }
 }
