+++ src/main/java/com/gradems/grademangementsystem/service/GradeServiceImpl.java
@@ -1,6 +1,7 @@
 package com.gradems.grademangementsystem.service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 import com.gradems.grademangementsystem.entity.Course;
 import com.gradems.grademangementsystem.entity.Grade;
 import com.gradems.grademangementsystem.entity.Student;
+import com.gradems.grademangementsystem.exception.NotFoundException;
 import com.gradems.grademangementsystem.repository.GradeRepository;
 
 import lombok.AllArgsConstructor;
@@ -21,12 +23,13 @@ public class GradeServiceImpl implements GradeService {
     CourseService courseService;
 
     @Override
-    public Grade getGrade(UUID studentId, UUID CourseId) {
-        return gradeRepository.findByStudentIdAndCourseId(studentId, CourseId).get();
+    public Grade getGrade(UUID studentId, UUID courseId) {
+        Optional<Grade> grade = gradeRepository.findByStudentIdAndCourseId(studentId, courseId);
+        return unwrapGrade(grade, studentId, courseId);
     }
 
     @Override
-    public Grade savGrade(Grade grade, UUID studentId, UUID courseId) {
+    public Grade saveGrade(Grade grade, UUID studentId, UUID courseId) {
         Student student = studentService.getStudent(studentId);
         Course course = courseService.getCourse(courseId);
         grade.setCourse(course);
@@ -36,14 +39,16 @@ public class GradeServiceImpl implements GradeService {
 
     @Override
     public Grade updatGrade(String score, UUID studentId, UUID courseId) { 
-        Grade grade = getGrade(studentId, courseId); //TODO Test null exception.
+        Grade grade = getGrade(studentId, courseId);
         grade.setScore(score);
         return gradeRepository.save(grade);
     }
 
     @Override
     public void deleteGrad(UUID studentId, UUID courseId) {
-        gradeRepository.deleteByStudentIdAndCourseId(studentId, courseId);
+        if(getGrade(studentId, courseId) != null){
+            gradeRepository.deleteByStudentIdAndCourseId(studentId, courseId);
+        }
     }
 
     @Override
@@ -59,6 +64,14 @@ public class GradeServiceImpl implements GradeService {
     @Override
     public List<Grade> getAllGrades() {
         return (List<Grade>)gradeRepository.findAll();
+    }
+
+    static Grade unwrapGrade(Optional<Grade> grade, UUID sid, UUID cid) {
+        if(grade.isPresent()){
+            return grade.get();
+        } else {
+            throw new NotFoundException(sid, cid);
+        }
     }
     
 }
